backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
packaging==26.0
setuptools==82.0.0
sqlparse==0.5.5
wheel==0.46.3
